backend/requirements.txt
fastapi
uvicorn
ultralytics
pillow
requests



